Refactored utilities file, extracted several methods out of ticker component. Extracted ticker price history graph into a separate component that handles it's own state. Price history graph can now make queries to the backend api with a start and end time that triggers whenever a user zooms on the graph using the buttons or dragging across the X-axis of the graph.
