+++ source/application/utilities/index.js
@@ -1,0 +1,7 @@
+import colors from './colors'
+import numbers from './numbers'
+
+export default {
+    colors,
+    numbers
+}
